Move redirected headers test to jvm package
